+++ routes/admin.php
@@ -91,6 +91,7 @@ Route::group(['prefix' => 'nodes', 'middleware' => 'admin.superuser'], function 
     Route::get('/view/{node:id}/allocation', [Admin\Nodes\NodeViewController::class, 'allocations'])->name('admin.nodes.view.allocation');
     Route::get('/view/{node:id}/servers', [Admin\Nodes\NodeViewController::class, 'servers'])->name('admin.nodes.view.servers');
     Route::get('/view/{node:id}/system-information', Admin\Nodes\SystemInformationController::class);
+    Route::get('/view/{node:id}/live-usage', Admin\Nodes\NodeSystemUsageController::class);
 
     Route::post('/new', [Admin\NodesController::class, 'store']);
     Route::post('/view/{node:id}/allocation', [Admin\NodesController::class, 'createAllocation']);